Refactor external source files to live in app/
