+++ package.json
@@ -109,12 +109,9 @@
       "stylesheets/*.css",
       "!js/register.js",
       "!js/views/standalone_registration_view.js",
+      "app/*",
       "preload.js",
-      "window_state.js",
-      "autoupdate.js",
-      "locale.js",
       "main.js",
-      "menu.js",
       "audio/**",
       "images/**",
       "fonts/*",
